sustainabull_backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
Django==5.0.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpointer==2.1
MarkupSafe==3.0.2
packaging==23.1
# psycopg2==2.9.9
psycopg2-binary==2.9.9
gunicorn==20.1.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
regex==2023.6.3
requests==2.31.0
sqlparse==0.5.0
tiktoken==0.4.0
tqdm==4.64.1
typing_extensions==4.11.0
urllib3==1.26.20
Werkzeug==3.1.3
